+++ frame/src/main/java/com/dreampany/frame/data/source/repository/StateRepository.java
@@ -34,7 +34,7 @@ public class StateRepository extends Repository<Long, State> implements StateDat
     }
 
     @Override
-    public int getCount(long id, String type, String subtype) {
+    public long getCount(long id, String type, String subtype) {
         return room.getCount(id, type, subtype);
     }
 
@@ -49,22 +49,22 @@ public class StateRepository extends Repository<Long, State> implements StateDat
     }
 
     @Override
-    public int getCount(String type, String subtype, String state) {
+    public long getCount(String type, String subtype, String state) {
         return room.getCount(type, subtype, state);
     }
 
     @Override
-    public int getCount(long id, String type, String subtype, String state) {
+    public long getCount(long id, String type, String subtype, String state) {
         return room.getCount(id, type, subtype, state);
     }
 
     @Override
-    public Maybe<Integer> getCountRx(String type, String subtype, String state) {
+    public Maybe<Long> getCountRx(String type, String subtype, String state) {
         return room.getCountRx(type, subtype, state);
     }
 
     @Override
-    public Maybe<Integer> getCountRx(long id, String type, String subtype, String state) {
+    public Maybe<Long> getCountRx(long id, String type, String subtype, String state) {
         return room.getCountRx(id, type, subtype, state);
     }
 
@@ -84,12 +84,12 @@ public class StateRepository extends Repository<Long, State> implements StateDat
     }
 
     @Override
-    public int getCount() {
+    public long getCount() {
         return room.getCount();
     }
 
     @Override
-    public Maybe<Integer> getCountRx() {
+    public Maybe<Long> getCountRx() {
         return room.getCountRx();
     }
 
@@ -264,12 +264,12 @@ public class StateRepository extends Repository<Long, State> implements StateDat
     }
 
     @Override
-    public List<State> getItems(int limit) {
+    public List<State> getItems(long limit) {
         return room.getItems(limit);
     }
 
     @Override
-    public Maybe<List<State>> getItemsRx(int limit) {
+    public Maybe<List<State>> getItemsRx(long limit) {
         return room.getItemsRx(limit);
     }
 }
