+++ frame/src/main/java/com/dreampany/frame/data/source/room/StoreRoomDataSource.java
@@ -64,12 +64,12 @@ public class StoreRoomDataSource implements StoreDataSource {
     }
 
     @Override
-    public int getCount() {
+    public long getCount() {
         return dao.getCount();
     }
 
     @Override
-    public Maybe<Integer> getCountRx() {
+    public Maybe<Long> getCountRx() {
         return dao.getCountRx();
     }
 
@@ -144,12 +144,12 @@ public class StoreRoomDataSource implements StoreDataSource {
     }
 
     @Override
-    public List<Store> getItems(int limit) {
+    public List<Store> getItems(long limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Store>> getItemsRx(int limit) {
+    public Maybe<List<Store>> getItemsRx(long limit) {
         return null;
     }
 }
