+++ lca/src/main/java/com/dreampany/lca/data/source/room/IcoRoomDataSource.java
@@ -121,12 +121,12 @@ public class IcoRoomDataSource implements IcoDataSource {
     }
 
     @Override
-    public List<Ico> getItems(int limit) {
+    public List<Ico> getItems(long limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Ico>> getItemsRx(int limit) {
+    public Maybe<List<Ico>> getItemsRx(long limit) {
         return null;
     }
 
@@ -136,32 +136,32 @@ public class IcoRoomDataSource implements IcoDataSource {
     }
 
     @Override
-    public List<Ico> getLiveItems(int limit) {
+    public List<Ico> getLiveItems(long limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Ico>> getLiveItemsRx(int limit) {
+    public Maybe<List<Ico>> getLiveItemsRx(long limit) {
         return dao.getItemsRx(IcoStatus.LIVE.name(), limit);
     }
 
     @Override
-    public List<Ico> getUpcomingItems(int limit) {
+    public List<Ico> getUpcomingItems(long limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Ico>> getUpcomingItemsRx(int limit) {
+    public Maybe<List<Ico>> getUpcomingItemsRx(long limit) {
         return dao.getItemsRx(IcoStatus.UPCOMING.name(), limit);
     }
 
     @Override
-    public List<Ico> getFinishedItems(int limit) {
+    public List<Ico> getFinishedItems(long limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Ico>> getFinishedItemsRx(int limit) {
+    public Maybe<List<Ico>> getFinishedItemsRx(long limit) {
         return dao.getItemsRx(IcoStatus.FINISHED.name(), limit);
     }
