+++ lca/src/main/java/com/dreampany/lca/vm/FinishedIcoViewModel.java
@@ -37,7 +37,7 @@ public class FinishedIcoViewModel
         extends BaseViewModel<Ico, IcoItem, UiTask<Ico>>
         implements NetworkManager.Callback {
 
-    private static final int LIMIT = Constants.Limit.ICO;
+    private static final long LIMIT = Constants.Limit.ICO;
     private static final long initialDelay = 0L;
     private static final long period = Constants.Time.INSTANCE.getIcoPeriod();
     private static final int RETRY_COUNT = 3;
