+++ word/src/main/java/com/dreampany/word/data/source/assets/WordAssetsDataSource.java
@@ -146,12 +146,12 @@ public class WordAssetsDataSource implements WordDataSource {
     }
 
     @Override
-    public List<Word> getItems(int limit) {
+    public List<Word> getItems(long limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Word>> getItemsRx(int limit) {
+    public Maybe<List<Word>> getItemsRx(long limit) {
         return null;
     }
