+++ word/src/main/java/com/dreampany/word/data/source/firestore/WordFirestoreDataSource.java
@@ -142,12 +142,12 @@ public class WordFirestoreDataSource implements WordDataSource {
     }
 
     @Override
-    public List<Word> getItems(int limit) {
+    public List<Word> getItems(long limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Word>> getItemsRx(int limit) {
+    public Maybe<List<Word>> getItemsRx(long limit) {
         return null;
     }
